+++ test/cli.js
@@ -1,7 +1,9 @@
 var bag = require('bagofcli'),
-  buster = require('buster'),
+  buster = require('buster-node'),
   cli = require('../lib/cli'),
-  DataGen = new require('../lib/datagen');
+  DataGen = new require('../lib/datagen'),
+  referee = require('referee'),
+  assert = referee.assert;
 
 buster.testCase('cli - exec', {
   'should contain commands with actions': function (done) {