+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.10-pre
-*
+* Change test lib to buster-node + referee
 
 ### 0.0.9
 * Add max concurrent workers option to gen command